04-유상민/비트컴퓨터_프로젝트_발표_streamlit/requirements.txt
langchain>=0.0.200
openai>=0.27.0
langchain-community>=0.0.1
langchain-teddynote>=0.0.1
beautifulsoup4>=4.12.2
pypdf>=3.7.0
pymupdf>=1.21.1
pypdfium2>=3.20.0
pdfminer.six>=20221105
pdfplumber>=0.5.28
unstructured>=0.7.8
openpyxl>=3.0.10
bs4
python-dotenv
validators>=0.20.0
jsonschema>=4.19.0
requests>=2.28.0
langchain_openai

transformers
sentence-transformers
torch

chromadb==0.4.15
python-dotenv